Debugged Dijkstra_4 algorithm
Found minor bug in Dijkstra_4 algorithm and eliminated it.
The bug was caused due to ParentIndex having value -1
sometimes before entering the swapping loop (Read the comment added)
and hence, disrupting the heap property.

While this fixed one observed instance of wrong path, [s = 1346748279,
d = 7079392136], along with many unobserved instances of the same
bug, the first bad path is still not fixed.
It is now unclear whether the first found bad-path (on input
[s = 7044039703, d = 7054927767]) is caused due to buggy road network
data or a problem in the actual pathfinding.
There seems to be no way of checking either since Dijkstra_3
takes forever to run on the same.